requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coloredlogs==15.0.1
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
Django==4.0.3
django-filter==21.1
djangorestframework==3.13.1
entrypoints==0.4
executing==0.8.3
frozenlist==1.3.0
humanfriendly==10.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.4.0
ipykernel==6.9.2
ipython==8.1.1
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.18.1
Jinja2==3.1.0
joblib==1.1.0
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.0
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
multidict==6.0.2
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.2.0
nest-asyncio==1.5.4
notebook==6.4.10
numpy==1.22.3
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.13.1
prompt-toolkit==3.0.28
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.11.2
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
requests==2.27.1
scikit-learn==1.0.2
scipy==1.8.0
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
soupsieve==2.3.1
sqlparse==0.4.2
stack-data==0.2.0
terminado==0.13.3
testpath==0.6.0
threadpoolctl==3.1.0
tornado==6.1
traitlets==5.1.1
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.6.0
yarl==1.7.2
zipp==3.7.0
